+++ apps/ledger-live-desktop/src/renderer/families/solana/MemoValueField.tsx
@@ -9,6 +9,7 @@ import {
   SolanaAccount,
 } from "@ledgerhq/live-common/families/solana/types";
 import { useFeature } from "@ledgerhq/live-common/featureFlags/index";
+import { SolanaRecipientMemoIsRequired } from "@ledgerhq/live-common/errors";
 import MemoTagField from "~/newArch/features/MemoTag/components/MemoTagField";
 
 type Props = {
@@ -43,6 +44,7 @@ const MemoValueField = ({ onChange, account, transaction, status, autoFocus }: P
   );
 
   const InputField = lldMemoTag?.enabled ? MemoTagField : Input;
+  const isRecipientMemoRequired = status?.errors?.memo instanceof SolanaRecipientMemoIsRequired;
 
   return transaction.model.kind === "transfer" || transaction.model.kind === "token.transfer" ? (
     <InputField
@@ -50,7 +52,11 @@ const MemoValueField = ({ onChange, account, transaction, status, autoFocus }: P
       error={status.errors.memo}
       value={transaction.model.uiState.memo || ""}
       onChange={onMemoValueChange}
-      placeholder={t("families.solana.memoPlaceholder")}
+      placeholder={t(
+        isRecipientMemoRequired
+          ? "families.solana.requiredMemoPlaceholder"
+          : "families.solana.memoPlaceholder",
+      )}
       autoFocus={autoFocus}
     />
   ) : null;