+++ apps/ledger-live-mobile/e2e/specs/speculos/send/sendValidAddress/sendValidAddressATOM.spec.ts
@@ -3,4 +3,4 @@ import { Transaction } from "@ledgerhq/live-common/e2e/models/Transaction";
 import { runSendValidAddressTest } from "../send";
 
 const transaction = new Transaction(Account.ATOM_1, Account.ATOM_2, "0.00001", undefined, "123456");
-runSendValidAddressTest(transaction, "B2CQA-2720");
+runSendValidAddressTest(transaction, ["B2CQA-2720"]);