+++ apps/ledger-live-mobile/src/screens/OperationDetails/Content.tsx
@@ -379,7 +379,7 @@ export default function Content({
         >
           {operation.fee ? (
             <View style={styles.feeValueContainer}>
-              <LText style={sectionStyles.value} semiBold>
+              <LText style={sectionStyles.value} semiBold testID="operationDetails-fees">
                 <CurrencyUnitValue showCode unit={feeUnit} value={operation.fee} />
               </LText>
               <LText style={styles.feeCounterValue} color="smoke" semiBold>