+++ libs/coin-modules/coin-tron/src/bridge/synchronization.integ.test.ts
@@ -1,11 +1,11 @@
-import BigNumber from "bignumber.js";
-import { firstValueFrom, reduce } from "rxjs";
 import { getCryptoCurrencyById } from "@ledgerhq/cryptoassets/index";
 import { Account, AccountBridge, SyncConfig, TransactionCommon } from "@ledgerhq/types-live";
+import BigNumber from "bignumber.js";
+import { firstValueFrom, reduce } from "rxjs";
 import { TronCoinConfig } from "../config";
-import { defaultTronResources } from "../logic/utils";
 import { Transaction, TronAccount } from "../types";
 import { createBridges } from "./index";
+import { defaultTronResources } from "./utils";
 
 const tron = getCryptoCurrencyById("tron");
 const defaultSyncConfig = {
@@ -74,7 +74,7 @@ describe("Sync Accounts", () => {
     bridge = createBridges(signer, coinConfig);
   });
 
-  test("should always have tronResources", async () => {
+  it("should always have tronResources", async () => {
     const account = await syncAccount<Transaction, TronAccount>(
       bridge.accountBridge,
       dummyAccount,
@@ -84,23 +84,41 @@ describe("Sync Accounts", () => {
     expect(account.tronResources).toEqual(defaultTronResources);
   });
 
-  test.each([
-    "TL24LCps5FKwp3PoU1MvrYrwhi5LU1tHre",
-    "TAVrrARNdnjHgCGMQYeQV7hv4PSu7mVsMj",
-    "THAe4BNVxp293qgyQEqXEkHMpPcqtG73bi",
-    "TRqkRnAj6ceJFYAn2p1eE7aWrgBBwtdhS9",
-    "TUxd6v64YTWkfpFpNDdtgc5Ps4SfGxwizT",
-    "TY2ksFgpvb82TgGPwUSa7iseqPW5weYQyh",
-  ])(
-    "should always be sync without error for address %s",
-    async (accountId: string) => {
+  it(
+    "should always be sync without error",
+    async () => {
+      // GIVEN
+      const id = "TL24LCps5FKwp3PoU1MvrYrwhi5LU1tHre";
+
+      // WHEN
       const account = await syncAccount<Transaction, TronAccount>(bridge.accountBridge, {
         ...dummyAccount,
-        id: `js:2:tron:${accountId}:`,
-        freshAddress: accountId,
+        id: `js:2:tron:${id}:`,
+        freshAddress: id,
       });
 
-      expect(account.id).toEqual(`js:2:tron:${accountId}:`);
+      // THEN
+      expect(account.id).toEqual(`js:2:tron:${id}:`);
+      expect(account.freshAddress).toEqual("TL24LCps5FKwp3PoU1MvrYrwhi5LU1tHre");
+      expect(account.operations[account.operations.length - 1]).toEqual({
+        accountId: "js:2:tron:TL24LCps5FKwp3PoU1MvrYrwhi5LU1tHre:",
+        blockHash: null,
+        blockHeight: 24725965,
+        date: new Date("2020-11-04T14:36:33.000Z"),
+        extra: {},
+        fee: new BigNumber("0"),
+        hasFailed: false,
+        hash: "22f871f18d39b6c39e3c1495ba529169bee3fbefd59b504dac15becaff264920",
+        id: "js:2:tron:TL24LCps5FKwp3PoU1MvrYrwhi5LU1tHre:-22f871f18d39b6c39e3c1495ba529169bee3fbefd59b504dac15becaff264920-IN",
+        recipients: ["TL24LCps5FKwp3PoU1MvrYrwhi5LU1tHre"],
+        senders: ["TNDoSUL32A2KRqbEKXZQuPWgfBcA42sCwM"],
+        type: "IN",
+        value: new BigNumber("11234560"),
+      });
+      const separator = "%2F";
+      expect(account.subAccounts![0].id).toEqual(
+        `js:2:tron:TL24LCps5FKwp3PoU1MvrYrwhi5LU1tHre:+tron${separator}trc10${separator}1002000`,
+      );
     },
     15 * 1_000,
   );