+++ libs/ui/examples/webpack.js/CHANGELOG.md
@@ -1,5 +1,19 @@
 # webpack.js-example
 
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/react-ui@0.19.1
+
+## 0.1.16-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/react-ui@0.19.1-next.0
+
 ## 0.1.15
 
 ### Patch Changes